Updating List.h int to long

+++ pa6/List.h
@@ -12,7 +12,7 @@
 #define List_H_INCLUDE_
 
 // Exported types -------------------------------------------------------------
-typedef int ListElement;
+typedef long ListElement;
 
 class List{
